+++ pkg/pub_package_reader/test/package_archive_test.dart
@@ -993,4 +993,33 @@ dev_dependencies:
       );
     });
   });
+
+  group('hooks', () {
+    test('allowed', () {
+      expect(
+        checkHooks(Version.parse('3.6.0'), {
+          'hook/build.dart',
+          'lib/x.dart',
+        }),
+        isEmpty,
+      );
+    });
+
+    test('prevented', () {
+      expect(
+        checkHooks(Version.parse('3.6.0'), {
+          'hook/x.dart',
+        }).first.message,
+        contains('`hook/x.dart` is not allowed'),
+      );
+
+      expect(
+        checkHooks(Version.parse('3.5.0'), {
+          'hook/build.dart',
+        }).first.message,
+        contains(
+            '`hook/build.dart` is allowed only with a minimum SDK constraint of'),
+      );
+    });
+  });
 }